requirements.txt
Babel==2.3.4
Django==1.10.3
django-filter==0.15.3
django-modeltranslation==0.12
django-phonenumber-field==1.1.0
django-translation-client==0.3.1
djangorestframework==3.5.3
Markdown==2.6.7
phonenumberslite==7.7.3
psycopg2==2.6.2
pytz==2016.7
requests==2.11.1
